% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_input_files.R
\name{load_DIANN_protein_input}
\alias{load_DIANN_protein_input}
\title{Load a DIANN pg matrix}
\usage{
load_DIANN_protein_input(path_to_input_protein, remove_path = TRUE)
}
\arguments{
\item{path_to_input_protein}{Character. Path to the DIANN \code{.tsv}
Protein Group matrix file (pg_matrix).}
}
\value{
A \code{tibble} in long format with one row per protein group-run
combination and the following columns:
\describe{
\item{condition}{Experimental condition (\code{NA} for DIANN pg_matrix outputs).}
\item{filename}{Raw file name (corresponding headers in the pg_matrix).}
\item{replicate}{Replicate label (\code{NA} for DIANN pg_matrix outputs).}
\item{pg_informations}{FASTA headers (\code{NA} for DIANN pg_matrix outputs}).}
\item{protein_group_accessions}{Protein accessions
(\code{Protein.Group}).}
\item{nr_precursors_used_for_quantification}{Number of precursors used
for quantification (\code{NA} for DIANN pg_matrix outputs).}
\item{nr_proteotypic_peptide_seq_for_id}{Number of proteotypic peptide
sequences used for identification (\code{N.Proteotypic.Sequences}).}
\item{quantity}{Raw protein group quantity (from corresponding columns in the DIANN pg_matrix).}
\item{log2_quan}{Log2-transformed quantity.}
}
}
\description{
Reads a tab-separated DIANN Protein Group matrix, verifies that all
required columns are present, selects and renames relevant columns, and
adds a log2-transformed quantity column.
}
